Change content on session timeout page to 20 minutes. Rename exit directory to exit-points
