+++ steps/exit-points/session-timeout/content.en.json
@@ -1,7 +1,7 @@
 {
   "titleHead": "Session timed out - Appeal a benefit decision - GOV.UK",
   "title": "Sorry, you'll have to start again",
-  "ends": "Your session automatically ends if you don't use this service for 30 minutes.",
+  "ends": "Your session automatically ends if you don't use this service for 20 minutes.",
   "detailsDeleted": "Your details have been deleted. We do this for your security.",
   "start": "Start again"
 }
